feat: Add event listener for task addition button
Implemented JavaScript code to attach a click event listener to the element with the id 'push'. When the button is clicked, it checks if the value of the input field inside the element with the id 'newtask' is empty. If so, it displays an alert instructing the user to enter a task.